+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Changelog
 - Add multi-tenant support via an index-prefix domain mapping file
 - PSL overrides so that services like AWS are correctly identified
 - Additional improvements to report type detection
+- Fix webhook timeout parsing (PR #623)
+- Output to STDOUT when the new general config boolean `silent` is set to `False` (Close #614)
 - Additional services added to `base_reverse_dns_map.csv`
 
 8.18.9